Rename database package to core
