+++ service/cornac/core/__init__.py
@@ -1,0 +1,4 @@
+#
+# Core package shares objects and functions used by every pieces of cornac: web
+# handlers, CLI scripts, background workers.
+#
