--- service/cornac/database/__init__.py
@@ -1,12 +1,0 @@
-from contextlib import contextmanager
-
-import psycopg2
-
-
-@contextmanager
-def connect(connstring):
-    conn = psycopg2.connect(connstring)
-    try:
-        yield conn
-    finally:
-        conn.close()
